replaced jopt-simple in KafkaDataStoreLogViewer with jcommander
