+++ geomesa-kafka/pom.xml
@@ -18,9 +18,6 @@
     <properties>
         <kafka.version>0.8.2.1</kafka.version>
         <commons-lang3.version>3.3.2</commons-lang3.version>
-
-        <!--Match version used by Kafka -->
-        <jopt.simple.version>3.2</jopt.simple.version>
     </properties>
 
     <dependencyManagement>
@@ -38,10 +35,6 @@
                     <exclusion>
                         <groupId>org.xerial.snappy</groupId>
                         <artifactId>snappy-java</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>net.sf.jopt-simple</groupId>
-                        <artifactId>jopt-simple</artifactId>
                     </exclusion>
                 </exclusions>
             </dependency>
@@ -75,12 +68,6 @@
                         <artifactId>jopt-simple</artifactId>
                     </exclusion>
                 </exclusions>
-            </dependency>
-            <dependency>
-                <groupId>net.sf.jopt-simple</groupId>
-                <artifactId>jopt-simple</artifactId>
-                <version>3.2</version>
-                <scope>provided</scope>
             </dependency>
         </dependencies>
     </dependencyManagement>
